Warning fixes (#887)
* Fix some code warnings from MSVC and clang_tidy, main change is changing broker_state to a scoped enumeration.

* release docs updates

* small tweak in cmake

* actually make the shared-cxx-tests work.  They were failing before and this was only detected in the daily builds

* turn off the examples on mingw test on travis

* fix ci command file

* fix some instances of changed cmake macro

* add tsan suppression file and fix an incorrect modification to the python build

* update some warnings on google benchmarks and more updates to the changelog

* a few more warning fixes and changelog updates

* codacy fixes and remove last remnants of HELICS_DISABLE_STATIC_LIB_INSTALL

* fix a warning from travis about unused variable and a sign mismatch in MessageFederateExport

* Apply suggestions from code review

.cmake for files and a few other grammar corrections

+++ benchmarks/helics/echoBenchmarks.cpp
@@ -184,7 +184,7 @@ static void BM_echo_singleCore (benchmark::State &state)
         state.PauseTiming ();
 
         int feds = static_cast<int> (state.range (0));
-        gmlc::concurrency::Barrier brr (feds + 1);
+        gmlc::concurrency::Barrier brr (static_cast<size_t>(feds) + 1);
         auto wcore = helics::CoreFactory::create (core_type::INPROC, std::string ("--autobroker --federates=") +
                                                                        std::to_string (feds + 1));
         EchoHub hub;
@@ -230,7 +230,7 @@ static void BM_echo_multiCore (benchmark::State &state, core_type cType)
         state.PauseTiming ();
 
         int feds = static_cast<int> (state.range (0));
-        gmlc::concurrency::Barrier brr (feds + 1);
+        gmlc::concurrency::Barrier brr (static_cast<size_t>(feds) + 1);
 
         auto broker = helics::BrokerFactory::create (cType, "brokerb",
                                                      std::string ("--federates=") + std::to_string (feds + 1));
